test: run the Gemma 4 multimodal test in its own mobile group
